refactor: move utils/helpers around, split into files
