+++ src/pages/popup/components/PayWebsiteForm.tsx
@@ -2,8 +2,8 @@ import React from 'react';
 import { Button } from '@/pages/shared/components/ui/Button';
 import { InputAmount } from '@/pages/shared/components/InputAmount';
 import { FadeInOut } from '@/pages/shared/components/FadeInOut';
-import { roundWithPrecision } from '@/pages/shared/lib/utils';
-import { cn, type ErrorWithKeyLike } from '@/shared/helpers';
+import { roundWithPrecision, cn } from '@/pages/shared/lib/utils';
+import type { ErrorWithKeyLike } from '@/shared/helpers';
 import { useMessage, useTranslation } from '@/popup/lib/context';
 import { usePopupState } from '@/popup/lib/store';
