+++ src/pages/progress-connect/components/Countdown.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { cn } from '@/shared/helpers';
+import { cn } from '@/pages/shared/lib/utils';
 
 export function Countdown({
   expiresAt,
